#166 fix: csAssembler 함수 -> DTO, Entity 이전
