+++ src/main/kotlin/com/psr/psr/cs/dto/response/FaqRes.kt
@@ -1,6 +1,7 @@
 package com.psr.psr.cs.dto.response
 
 import com.fasterxml.jackson.annotation.JsonInclude
+import com.psr.psr.cs.entity.Faq
 
 data class FaqRes (
     val faqId: Long,
@@ -8,4 +9,10 @@ data class FaqRes (
     val title: String,
     @JsonInclude(JsonInclude.Include.NON_NULL)
     val content: String?= null,
-)
+){
+    companion object{
+        fun toFaqRes(faq: Faq): FaqRes {
+            return FaqRes(faq.id, faq.type.value, faq.title, faq.content)
+        }
+    }
+}
